Break up writers.cpp into individual writers
