+++ include/morphio/vector_types.h
@@ -15,12 +15,15 @@ using range = gsl::span<T>;
 using floatType = double;
 constexpr floatType epsilon = 1e-6;
 constexpr floatType PI = M_PI;
+constexpr int FLOAT_PRECISION_PRINT = 17;
+
 #else
 /** Type of float to use. Can be double or float depending on MORPHIO_USE_DOUBLE */
 using floatType = float;
 /** A really small value that is used to measure how close are two values */
 constexpr floatType epsilon = 1e-6F;
 constexpr floatType PI = static_cast<floatType>(M_PI);
+constexpr int FLOAT_PRECISION_PRINT = 9;
 #endif
 
 /** An array of size 3 for x,y,z coordinates */
