+++ src/mut/morphology.cpp
@@ -3,10 +3,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 #include <algorithm>  // std::find_if
-#include <cassert>
 #include <cctype>    // std::tolower
 #include <iterator>  // std::back_inserter
-#include <sstream>
 #include <string>
 
 #include <morphio/endoplasmic_reticulum.h>
